+++ pkg/volume/metrics_du_test.go
@@ -1,6 +1,3 @@
-//go:build linux
-// +build linux
-
 /*
 Copyright 2015 The Kubernetes Authors.
 
@@ -25,21 +22,10 @@ import (
 	"path/filepath"
 	"testing"
 
-	"golang.org/x/sys/unix"
 	utiltesting "k8s.io/client-go/util/testing"
 	. "k8s.io/kubernetes/pkg/volume"
 	volumetest "k8s.io/kubernetes/pkg/volume/testing"
 )
-
-func getExpectedBlockSize(path string) int64 {
-	statfs := &unix.Statfs_t{}
-	err := unix.Statfs(path, statfs)
-	if err != nil {
-		return 0
-	}
-
-	return int64(statfs.Bsize)
-}
 
 // TestMetricsDuGetCapacity tests that MetricsDu can read disk usage
 // for path
