+++ pkg/volume/metrics_du_windows_test.go
@@ -1,0 +1,32 @@
+//go:build windows
+// +build windows
+
+/*
+Copyright 2023 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package volume_test
+
+import (
+	"k8s.io/kubernetes/pkg/volume/util/fs"
+)
+
+func getExpectedBlockSize(path string) int64 {
+	usage, err := fs.DiskUsage(path)
+	if err != nil {
+		return 0
+	}
+	return usage.Bytes
+}
