+++ pkg/volume/util/hostutil/hostutil_linux_test.go
@@ -20,15 +20,10 @@ limitations under the License.
 package hostutil
 
 import (
-	"fmt"
 	"io/ioutil"
-	"net"
 	"os"
 	"path/filepath"
-	"strings"
 	"testing"
-
-	"k8s.io/utils/exec"
 )
 
 func TestIsSharedSuccess(t *testing.T) {
@@ -195,127 +190,6 @@ func TestGetSELinuxSupport(t *testing.T) {
 			t.Errorf("Test %s failed: expected %v, got %v", test.name, test.expectedResult, out)
 		}
 	}
-}
-
-func createSocketFile(socketDir string) (string, error) {
-	testSocketFile := filepath.Join(socketDir, "mt.sock")
-
-	// Switch to volume path and create the socket file
-	// socket file can not have length of more than 108 character
-	// and hence we must use relative path
-	oldDir, _ := os.Getwd()
-
-	err := os.Chdir(socketDir)
-	if err != nil {
-		return "", err
-	}
-	defer func() {
-		os.Chdir(oldDir)
-	}()
-	_, socketCreateError := net.Listen("unix", "mt.sock")
-	return testSocketFile, socketCreateError
-}
-
-func TestGetFileType(t *testing.T) {
-	hu := NewHostUtil()
-
-	testCase := []struct {
-		name         string
-		expectedType FileType
-		setUp        func() (string, string, error)
-	}{
-		{
-			"Directory Test",
-			FileTypeDirectory,
-			func() (string, string, error) {
-				tempDir, err := ioutil.TempDir("", "test-get-filetype-")
-				return tempDir, tempDir, err
-			},
-		},
-		{
-			"File Test",
-			FileTypeFile,
-			func() (string, string, error) {
-				tempFile, err := ioutil.TempFile("", "test-get-filetype")
-				if err != nil {
-					return "", "", err
-				}
-				tempFile.Close()
-				return tempFile.Name(), tempFile.Name(), nil
-			},
-		},
-		{
-			"Socket Test",
-			FileTypeSocket,
-			func() (string, string, error) {
-				tempDir, err := ioutil.TempDir("", "test-get-filetype-")
-				if err != nil {
-					return "", "", err
-				}
-				tempSocketFile, err := createSocketFile(tempDir)
-				return tempSocketFile, tempDir, err
-			},
-		},
-		{
-			"Block Device Test",
-			FileTypeBlockDev,
-			func() (string, string, error) {
-				tempDir, err := ioutil.TempDir("", "test-get-filetype-")
-				if err != nil {
-					return "", "", err
-				}
-
-				tempBlockFile := filepath.Join(tempDir, "test_blk_dev")
-				outputBytes, err := exec.New().Command("mknod", tempBlockFile, "b", "89", "1").CombinedOutput()
-				if err != nil {
-					err = fmt.Errorf("%v: %s ", err, outputBytes)
-				}
-				return tempBlockFile, tempDir, err
-			},
-		},
-		{
-			"Character Device Test",
-			FileTypeCharDev,
-			func() (string, string, error) {
-				tempDir, err := ioutil.TempDir("", "test-get-filetype-")
-				if err != nil {
-					return "", "", err
-				}
-
-				tempCharFile := filepath.Join(tempDir, "test_char_dev")
-				outputBytes, err := exec.New().Command("mknod", tempCharFile, "c", "89", "1").CombinedOutput()
-				if err != nil {
-					err = fmt.Errorf("%v: %s ", err, outputBytes)
-				}
-				return tempCharFile, tempDir, err
-			},
-		},
-	}
-
-	for idx, tc := range testCase {
-		path, cleanUpPath, err := tc.setUp()
-		defer os.RemoveAll(cleanUpPath) // RemoveAll can deal with a empty path ""
-		if err != nil {
-			// Locally passed, but upstream CI is not friendly to create such device files
-			// Leave "Operation not permitted" out, which can be covered in an e2e test
-			if isOperationNotPermittedError(err) {
-				continue
-			}
-			t.Fatalf("[%d-%s] unexpected error : %v", idx, tc.name, err)
-		}
-
-		fileType, err := hu.GetFileType(path)
-		if err != nil {
-			t.Fatalf("[%d-%s] unexpected error : %v", idx, tc.name, err)
-		}
-		if fileType != tc.expectedType {
-			t.Fatalf("[%d-%s] expected %s, but got %s", idx, tc.name, tc.expectedType, fileType)
-		}
-	}
-}
-
-func isOperationNotPermittedError(err error) bool {
-	return strings.Contains(err.Error(), "Operation not permitted")
 }
 
 func writeFile(content string) (string, string, error) {
